api/requirements.txt
fastapi
uvicorn
Pillow
numpy
opencv-python-headless
